+++ src/de/uxnr/amf/v3/type/Integer.java
@@ -7,18 +7,18 @@ import de.uxnr.amf.AMF_Context;
 import de.uxnr.amf.v3.base.U29;
 
 public class Integer extends U29 {
-	public static final long MAX_VALUE = 268435455;
-	public static final long MIN_VALUE = -268435456;
-	
+	public static final int MAX_VALUE = 268435455;
+	public static final int MIN_VALUE = -268435456;
+
 	public Integer() {
 		this.signed = true;
 	}
-	
-	public Integer(long value) {
+
+	public Integer(int value) {
 		this.signed = true;
 		this.set(value);
 	}
-	
+
 	public Integer(AMF_Context context, DataInputStream input) throws IOException {
 		this.signed = true;
 		this.read(context, input);
